split collections, realm_object, lists
