+++ test/list_test.dart
@@ -264,4 +264,223 @@ Future<void> main([List<String>? args]) async {
     expect(teams[0].players.length, 1);
     realm.close();
   });
+
+  test('Access deleted object collection', () {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    var team = Team("TeamOne");
+    realm.write(() => realm.add(team));
+    var teams = realm.all<Team>();
+    realm.write(() => realm.delete(team));
+    expect(() => team.players, throws<RealmException>("Accessing object of type Team which has been invalidated or deleted"));
+    realm.close();
+  });
+
+  test('Delete collection of deleted parent', () {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    var team = Team("TeamOne");
+    realm.write(() => realm.add(team));
+    var players = team.players;
+    realm.write(() => realm.delete(team));
+    expect(() => realm.write(() => realm.deleteMany(players)), throws<RealmException>("Access to invalidated Collection object"));
+    realm.close();
+  });
+
+  test('Realm deleteMany from list after realm is closed', () {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    //Create a team
+    final team = Team("Ferrari");
+    realm.write(() => realm.add(team));
+
+    //Add players to the team
+    final newPlayers = [
+      Person("Michael Schumacher"),
+      Person("Sebastian Vettel"),
+      Person("Kimi Räikkönen"),
+    ];
+    realm.write(() => team.players.addAll(newPlayers));
+
+    //Ensure team exists in realm
+    var teams = realm.all<Team>();
+    expect(teams.length, 1);
+
+    //Try to delete team players while realm is closed
+    final players = teams[0].players;
+    realm.close();
+    expect(
+        () => realm.write(() {
+              realm.deleteMany(players);
+            }),
+        throws<RealmException>());
+
+    //Ensure all persons still exists in realm
+    realm = Realm(config);
+    final allPersons = realm.all<Person>();
+    expect(allPersons.length, 3);
+    realm.close();
+  });
+
+  test('Get list length after deleting parent objects', () {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    var team = Team("TeamOne")..players.add(Person("Nikos"));
+    realm.write(() {
+      realm.add(team);
+      realm.delete(team);
+    });
+    expect(() => team.players.length, throws<RealmException>("Accessing object of type Team which has been invalidated or deleted"));
+
+    realm.close();
+  });
+
+  test('List isValid', () {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    realm.write(() {
+      realm.add(Team("Speed Team", players: [
+        Person("Michael Schumacher"),
+        Person("Sebastian Vettel"),
+        Person("Kimi Räikkönen"),
+      ]));
+    });
+
+    var teams = realm.all<Team>();
+
+    expect(teams, isNotNull);
+    expect(teams.length, 1);
+    final players = teams[0].players as RealmList<Person>;
+    expect(players.isValid, true);
+    realm.close();
+    expect(players.isValid, false);
+  });
+
+  test('Realm deleteMany from list referenced by two objects', () {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    //Create two teams
+    final teamOne = Team("Ferrari");
+    final teamTwo = Team("Maserati");
+    realm.write(() {
+      realm.add(teamOne);
+      realm.add(teamTwo);
+    });
+
+    //Create common players list for both teams
+    final newPlayers = [
+      Person("Michael Schumacher"),
+      Person("Sebastian Vettel"),
+      Person("Kimi Räikkönen"),
+    ];
+    realm.write(() {
+      teamOne.players.addAll(newPlayers);
+      teamTwo.players.addAll(newPlayers);
+    });
+
+    //Ensule teams exist in realm
+    var teams = realm.all<Team>();
+    expect(teams.length, 2);
+
+    //Delete all players in a team from realm
+    realm.write(() => realm.deleteMany(teams[0].players));
+
+    //Ensure all players are deleted from collection
+    expect(teams[0].players.length, 0);
+
+    //Reload all persons from realm and ensure they are deleted
+    final allPersons = realm.all<Person>();
+    expect(allPersons.length, 0);
+    realm.close();
+  });
+
+  test('RealmObject add with list properties', () {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    final team = Team("Ferrari")
+      ..players.addAll([Person("Michael"), Person("Kimi")])
+      ..scores.addAll([1, 2, 3]);
+
+    realm.write(() => realm.add(team));
+
+    final teams = realm.all<Team>();
+    expect(teams.length, 1);
+    expect(teams[0].name, "Ferrari");
+    expect(teams[0].players, isNotNull);
+    expect(teams[0].players.length, 2);
+    expect(teams[0].players[0].name, "Michael");
+    expect(teams[0].players[1].name, "Kimi");
+
+    expect(teams[0].scores.length, 3);
+    expect(teams[0].scores[0], 1);
+    expect(teams[0].scores[1], 2);
+    expect(teams[0].scores[2], 3);
+    realm.close();
+  });
+
+  test('Realm deleteMany from iterable', () {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    //Create two teams
+    final teamOne = Team("Team one");
+    final teamTwo = Team("Team two");
+    final teamThree = Team("Team three");
+    realm.write(() {
+      realm.add(teamOne);
+      realm.add(teamTwo);
+      realm.add(teamThree);
+    });
+
+    //Ensure the teams exist in realm
+    var teams = realm.all<Team>();
+    expect(teams.length, 3);
+
+    //Delete teams one and three from realm
+    realm.write(() => realm.deleteMany([teamOne, teamThree]));
+
+    //Ensure both teams are deleted and only teamTwo has left
+    expect(teams.length, 1);
+    expect(teams[0].name, teamTwo.name);
+    realm.close();
+  });
+
+  test('Realm deleteMany from realm list', () {
+    var config = Configuration([Team.schema, Person.schema]);
+    var realm = Realm(config);
+
+    //Create a team
+    final team = Team("Ferrari");
+    realm.write(() => realm.add(team));
+
+    //Add players to the team
+    final newPlayers = [
+      Person("Michael Schumacher"),
+      Person("Sebastian Vettel"),
+      Person("Kimi Räikkönen"),
+    ];
+    realm.write(() => team.players.addAll(newPlayers));
+
+    //Ensure the team exists in realm
+    var teams = realm.all<Team>();
+    expect(teams.length, 1);
+
+    //Delete team players
+    realm.write(() => realm.deleteMany(teams[0].players));
+
+    //Ensure players are deleted from collection
+    expect(teams[0].players.length, 0);
+
+    //Reload all persons from realm and ensure they are deleted
+    final allPersons = realm.all<Person>();
+    expect(allPersons.length, 0);
+    realm.close();
+  });
 }
